+++ BuildResidentialHPXML/measure.rb
@@ -1060,6 +1060,11 @@ class BuildResidentialHPXML < OpenStudio::Measure::ModelMeasure
     arg.setDescription("Whether the heating system is condensing or not. Only used for a non-electric #{HPXML::HVACTypeBoiler}.")
     args << arg
 
+    arg = OpenStudio::Measure::OSArgument::makeBoolArgument('heating_system_outdoor_reset_control', false)
+    arg.setDisplayName('Heating System: Outdoor Reset Control')
+    arg.setDescription("Whether the heating system has outdoor reset control. Only used for a #{HPXML::HVACTypeBoiler}.")
+    args << arg
+
     arg = OpenStudio::Measure::OSArgument::makeDoubleArgument('heating_system_airflow_defect_ratio', false)
     arg.setDisplayName('Heating System: Airflow Defect Ratio')
     arg.setDescription("The airflow defect ratio, defined as (InstalledAirflow - DesignAirflow) / DesignAirflow, of the heating system per ANSI/RESNET/ACCA Standard 310. A value of zero means no airflow defect. Applies only to #{HPXML::HVACTypeFurnace}.")
@@ -4373,6 +4378,12 @@ class HPXMLFile
       end
     end
 
+    if heating_system_type == HPXML::HVACTypeBoiler
+      if args[:heating_system_outdoor_reset_control].is_initialized
+        heating_system_outdoor_reset_control = args[:heating_system_outdoor_reset_control].get
+      end
+    end
+
     if args[:heating_system_airflow_defect_ratio].is_initialized
       if [HPXML::HVACTypeFurnace].include? heating_system_type
         airflow_defect_ratio = args[:heating_system_airflow_defect_ratio].get
@@ -4399,6 +4410,7 @@ class HPXMLFile
                               heating_efficiency_afue: heating_efficiency_afue,
                               heating_efficiency_percent: heating_efficiency_percent,
                               condensing_system: heating_system_is_condensing,
+                              boiler_reset_control: heating_system_outdoor_reset_control,
                               airflow_defect_ratio: airflow_defect_ratio,
                               is_shared_system: is_shared_system,
                               number_of_units_served: number_of_units_served,