+++ HPXMLtoOpenStudio/resources/hvac.rb
@@ -423,19 +423,7 @@ class HVAC
   def self.apply_boiler(model, runner, heating_system,
                         sequential_heat_load_fracs, control_zone)
     obj_name = Constants.ObjectNameBoiler
-    oat_reset_enabled = false
-    oat_high = nil
-    oat_low = nil
-    oat_hwst_high = nil
-    oat_hwst_low = nil
     design_temp = 180.0 # deg-F
-
-    if oat_reset_enabled
-      if oat_high.nil? || oat_low.nil? || oat_hwst_low.nil? || oat_hwst_high.nil?
-        runner.registerWarning('Boiler outdoor air temperature (OAT) reset is enabled but no setpoints were specified so OAT reset is being disabled.')
-        oat_reset_enabled = false
-      end
-    end
 
     # Plant Loop
     plant_loop = OpenStudio::Model::PlantLoop.new(model)
@@ -472,7 +460,7 @@ class HVAC
     boiler = OpenStudio::Model::BoilerHotWater.new(model)
     boiler.setName(obj_name)
     boiler.setFuelType(EPlus.fuel_type(heating_system.heating_system_fuel))
-    if heating_system.condensing_system
+    if heating_system.condensing_system && heating_system.heating_system_fuel != HPXML::FuelTypeElectricity
       # Convert Rated Efficiency at 80F and 1.0PLR where the performance curves are derived from to Design condition as input
       boiler_rated_HWRT = UnitConversions.convert(80.0 - 32.0, 'R', 'K')
       boiler_design_HWRT = UnitConversions.convert(design_temp - 20.0 - 32.0, 'R', 'K')
@@ -500,25 +488,25 @@ class HVAC
     boiler.additionalProperties.setFeature('HPXML_ID', heating_system.id) # Used by reporting measure
     set_pump_power_ems_program(model, pump_w, pump, boiler)
 
-    if oat_reset_enabled
+    if heating_system.boiler_reset_control
       setpoint_manager_oar = OpenStudio::Model::SetpointManagerOutdoorAirReset.new(model)
       setpoint_manager_oar.setName(obj_name + ' outdoor reset')
       setpoint_manager_oar.setControlVariable('Temperature')
-      setpoint_manager_oar.setSetpointatOutdoorLowTemperature(UnitConversions.convert(oat_hwst_low, 'F', 'C'))
-      setpoint_manager_oar.setOutdoorLowTemperature(UnitConversions.convert(oat_low, 'F', 'C'))
-      setpoint_manager_oar.setSetpointatOutdoorHighTemperature(UnitConversions.convert(oat_hwst_high, 'F', 'C'))
-      setpoint_manager_oar.setOutdoorHighTemperature(UnitConversions.convert(oat_high, 'F', 'C'))
+      setpoint_manager_oar.setOutdoorLowTemperature(UnitConversions.convert(heating_system.boiler_reset_low_oat, 'F', 'C'))
+      setpoint_manager_oar.setSetpointatOutdoorLowTemperature(UnitConversions.convert(heating_system.boiler_reset_setpoint_at_low_oat, 'F', 'C'))
+      setpoint_manager_oar.setOutdoorHighTemperature(UnitConversions.convert(heating_system.boiler_reset_high_oat, 'F', 'C'))
+      setpoint_manager_oar.setSetpointatOutdoorHighTemperature(UnitConversions.convert(heating_system.boiler_reset_setpoint_at_high_oat, 'F', 'C'))
       setpoint_manager_oar.addToNode(plant_loop.supplyOutletNode)
+    else
+      hydronic_heat_supply_setpoint = OpenStudio::Model::ScheduleConstant.new(model)
+      hydronic_heat_supply_setpoint.setName(obj_name + ' hydronic heat supply setpoint')
+      hydronic_heat_supply_setpoint.setValue(UnitConversions.convert(design_temp, 'F', 'C'))
+
+      setpoint_manager_scheduled = OpenStudio::Model::SetpointManagerScheduled.new(model, hydronic_heat_supply_setpoint)
+      setpoint_manager_scheduled.setName(obj_name + ' hydronic heat loop setpoint manager')
+      setpoint_manager_scheduled.setControlVariable('Temperature')
+      setpoint_manager_scheduled.addToNode(plant_loop.supplyOutletNode)
     end
-
-    hydronic_heat_supply_setpoint = OpenStudio::Model::ScheduleConstant.new(model)
-    hydronic_heat_supply_setpoint.setName(obj_name + ' hydronic heat supply setpoint')
-    hydronic_heat_supply_setpoint.setValue(UnitConversions.convert(design_temp, 'F', 'C'))
-
-    setpoint_manager_scheduled = OpenStudio::Model::SetpointManagerScheduled.new(model, hydronic_heat_supply_setpoint)
-    setpoint_manager_scheduled.setName(obj_name + ' hydronic heat loop setpoint manager')
-    setpoint_manager_scheduled.setControlVariable('Temperature')
-    setpoint_manager_scheduled.addToNode(plant_loop.supplyOutletNode)
 
     pipe_supply_bypass = OpenStudio::Model::PipeAdiabatic.new(model)
     plant_loop.addSupplyBranchForComponent(pipe_supply_bypass)