+++ HPXMLtoOpenStudio/tests/test_defaults.rb
@@ -1160,9 +1160,14 @@ class HPXMLtoOpenStudioDefaultsTest < MiniTest::Test
     hpxml.heating_systems[0].heating_efficiency_afue = 0.85
     hpxml.heating_systems[0].year_installed = 2010
     hpxml.heating_systems[0].condensing_system = true
+    hpxml.heating_systems[0].boiler_reset_control = true
+    hpxml.heating_systems[0].boiler_reset_low_oat = 1.0
+    hpxml.heating_systems[0].boiler_reset_setpoint_at_low_oat = 199.0
+    hpxml.heating_systems[0].boiler_reset_high_oat = 69.0
+    hpxml.heating_systems[0].boiler_reset_setpoint_at_high_oat = 99.0
     XMLHelper.write_file(hpxml.to_oga, @tmp_hpxml_path)
     hpxml_default = _test_measure()
-    _test_default_boiler_values(hpxml_default.heating_systems[0], 99.9, 12345, 0.85, true)
+    _test_default_boiler_values(hpxml_default.heating_systems[0], 99.9, 12345, 0.85, true, true, 1.0, 199.0, 69.0, 99.0)
 
     # Test defaults w/ in-unit boiler
     hpxml.heating_systems[0].electric_auxiliary_energy = nil
@@ -1170,9 +1175,16 @@ class HPXMLtoOpenStudioDefaultsTest < MiniTest::Test
     hpxml.heating_systems[0].heating_efficiency_afue = nil
     hpxml.heating_systems[0].year_installed = 2010
     hpxml.heating_systems[0].condensing_system = nil
+    hpxml.heating_systems[0].boiler_reset_control = nil
     XMLHelper.write_file(hpxml.to_oga, @tmp_hpxml_path)
     hpxml_default = _test_measure()
-    _test_default_boiler_values(hpxml_default.heating_systems[0], 170.0, nil, 0.797, false)
+    _test_default_boiler_values(hpxml_default.heating_systems[0], 170.0, nil, 0.797, false, false)
+
+    # Test defaults w/ in-unit boiler and outdoor reset
+    hpxml.heating_systems[0].boiler_reset_control = true
+    XMLHelper.write_file(hpxml.to_oga, @tmp_hpxml_path)
+    hpxml_default = _test_measure()
+    _test_default_boiler_values(hpxml_default.heating_systems[0], 170.0, nil, 0.797, false, false, 0.0, 180.0, 68.0, 95.0)
 
     # Test inputs not overridden by defaults (shared boiler)
     hpxml = _create_hpxml('base-bldgtype-multifamily-shared-boiler-only-baseboard.xml')
@@ -1181,18 +1193,20 @@ class HPXMLtoOpenStudioDefaultsTest < MiniTest::Test
     hpxml.heating_systems[0].heating_efficiency_afue = 0.85
     hpxml.heating_systems[0].year_installed = 1980
     hpxml.heating_systems[0].condensing_system = true
+    hpxml.heating_systems[0].boiler_reset_control = true
     XMLHelper.write_file(hpxml.to_oga, @tmp_hpxml_path)
     hpxml_default = _test_measure()
-    _test_default_boiler_values(hpxml_default.heating_systems[0], 99.9, nil, 0.85, true)
+    _test_default_boiler_values(hpxml_default.heating_systems[0], 99.9, nil, 0.85, true, true)
 
     # Test defaults w/ shared boiler
     hpxml.heating_systems[0].electric_auxiliary_energy = nil
     hpxml.heating_systems[0].heating_efficiency_afue = nil
     hpxml.heating_systems[0].year_installed = 1980
     hpxml.heating_systems[0].condensing_system = nil
+    hpxml.heating_systems[0].boiler_reset_control = nil
     XMLHelper.write_file(hpxml.to_oga, @tmp_hpxml_path)
     hpxml_default = _test_measure()
-    _test_default_boiler_values(hpxml_default.heating_systems[0], 220.0, nil, 0.723, false)
+    _test_default_boiler_values(hpxml_default.heating_systems[0], 220.0, nil, 0.723, false, false)
   end
 
   def test_stoves
@@ -3450,7 +3464,9 @@ class HPXMLtoOpenStudioDefaultsTest < MiniTest::Test
     end
   end
 
-  def _test_default_boiler_values(heating_system, eae, heating_capacity, heating_efficiency_afue, condensing_system)
+  def _test_default_boiler_values(heating_system, eae, heating_capacity, heating_efficiency_afue, condensing_system,
+                                  reset_control, reset_low_oat = nil, reset_setpoint_at_low_oat = nil,
+                                  reset_hight_oat = nil, reset_setpoint_at_high_oat = nil)
     assert_equal(eae, heating_system.electric_auxiliary_energy)
     if heating_capacity.nil?
       assert(heating_system.heating_capacity > 0)
@@ -3463,6 +3479,27 @@ class HPXMLtoOpenStudioDefaultsTest < MiniTest::Test
       assert_equal(heating_system.heating_efficiency_afue, heating_efficiency_afue)
     end
     assert_equal(heating_system.condensing_system, condensing_system)
+    assert_equal(heating_system.reset_control, reset_control)
+    if reset_low_oat.nil?
+      assert_nil(heating_system.boiler_reset_low_oat)
+    else
+      assert_equal(heating_system.boiler_reset_low_oat, reset_low_oat)
+    end
+    if reset_setpoint_at_low_oat.nil?
+      assert_nil(heating_system.boiler_setpoint_at_reset_low_oat)
+    else
+      assert_equal(heating_system.boiler_setpoint_at_reset_low_oat, reset_setpoint_at_low_oat)
+    end
+    if reset_high_oat.nil?
+      assert_nil(heating_system.boiler_reset_high_oat)
+    else
+      assert_equal(heating_system.boiler_reset_high_oat, reset_high_oat)
+    end
+    if reset_setpoint_at_high_oat.nil?
+      assert_nil(heating_system.boiler_reset_setpoint_at_high_oat)
+    else
+      assert_equal(heating_system.boiler_reset_setpoint_at_high_oat, reset_setpoint_at_high_oat)
+    end
   end
 
   def _test_default_stove_values(heating_system, fan_watts, heating_capacity, heating_efficiency_percent)